backend/requirements.txt
# Smart Medi Assistant Backend - Python Dependencies

# ============================================================================
# FASTAPI & WEB
# ============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
pydantic>=2.10.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# ============================================================================
# DATABASE
# ============================================================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver
alembic>=1.13.0        # Database migrations

# ============================================================================
# CACHE & SESSION STORAGE
# ============================================================================
redis>=5.0.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
werkzeug>=3.0.0

# ============================================================================
# EURI API & LLM
# ============================================================================
openai>=1.68.0         # OpenAI SDK (used for Euri API too)
tenacity>=9.0.0        # Retry logic with exponential backoff

# ============================================================================
# VECTOR DATABASE (FAISS - Local)
# ============================================================================
faiss-cpu>=1.7.0       # Use faiss-gpu if you have GPU
numpy>=1.24.0

# ============================================================================
# DOCUMENT PROCESSING
# ============================================================================
PyPDF2>=3.0.0          # PDF extraction
Pillow>=11.0.0         # Image processing
python-docx>=0.8.11    # Word document support

# ============================================================================
# TESTING
# ============================================================================
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=4.1.0
httpx>=0.28.0          # HTTPX client for testing

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
python-dotenv>=1.1.0
structlog>=24.0.0

# ============================================================================
# UTILITIES
# ============================================================================
python-dateutil>=2.8.2
requests>=2.31.0
validators>=0.22.0     # Input validation

# ============================================================================
# CODE QUALITY (Optional)
# ============================================================================
# black>=24.0.0         # Code formatter
# flake8>=7.0.0         # Linter
# mypy>=1.8.0           # Type checker
# isort>=5.13.0         # Import sorting
